fix(obs): format dock sources and match the libobs stub ABI
The repository format gate rejects the unformatted dock additions, and the
unit-test stub must return long long to match obs_data_get_int exactly.

+++ cpp/obs/test/moq-output-test.cpp
@@ -156,7 +156,7 @@ uint32_t obs_encoder_get_height(const obs_encoder_t *)
 	return 1080;
 }
 
-int64_t obs_data_get_int(obs_data_t *, const char *)
+long long obs_data_get_int(obs_data_t *, const char *)
 {
 	return 8000;
 }
